shipperctl: move commands to their own packages
